Yana branch

+++ app/src/main/java/com/company/metrix/data/model/User.kt
@@ -4,7 +4,7 @@ data class User(
     val id: Long,
     val name : String,
     val email : String,
-    val team_id: Int,
+    val team_id: Long,
     val position : String,
     val role : String
 //    val ratings: List<Estimation>,
